refactor: simplify component structure and imports.
