--- frontend/src/components/appRoutes/AppRoutes.tsx
@@ -1,14 +1,0 @@
-import { Routes, Route } from 'react-router-dom';
-import { Home } from '../homePage';
-import { NotFound } from '../notFound';
-
-const AppRoutes: React.FC = () => {
-	return (
-		<Routes>
-			<Route path="/" element={<Home />} />
-			<Route path="*" element={<NotFound />} />
-		</Routes>
-	);
-};
-
-export default AppRoutes;
