+++ frontend/src/components/buttonLink/index.tsx
@@ -1,1 +1,18 @@
-export { default as ButtonLink } from './ButtonLink';
+import { Link } from 'react-router-dom';
+import '../buttonLink/ButtonLink.css';
+
+interface ButtonLinkProps {
+  buttontext: string;
+  to: string;
+  id?: string;
+}
+
+const ButtonLink: React.FC<ButtonLinkProps> = ({ buttontext, to, id }) => {
+  return (
+    <Link to={to} id={id} className="Link" aria-label={buttontext}>
+      {buttontext}
+    </Link>
+  );
+};
+
+export default ButtonLink;
