+++ frontend/src/components/notFound/index.tsx
@@ -1,1 +1,13 @@
-export { default as NotFound } from './NotFound';
+import { Link } from 'react-router-dom';
+
+const NotFound: React.FC = () => {
+	return (
+		<div>
+			<h1>Página não encontrada</h1>
+			<p>Desculpe, a página que você está tentando acessar não existe.</p>
+			<Link to="/">Voltar para a página inicial</Link>
+		</div>
+	);
+};
+
+export default NotFound;
